+++ FlutterWave.Core/Models/Services/Foundations/ExternalFlutterWave/ExternalTransactions/ExternalCreateRefundResponse.cs
@@ -32,7 +32,7 @@ namespace FlutterWave.Core.Models.Services.Foundations.ExternalFlutterWave.Exter
             public int WalletId { get; set; }
 
             [JsonProperty("amount_refunded")]
-            public int AmountRefunded { get; set; }
+            public decimal AmountRefunded { get; set; }
 
             [JsonProperty("status")]
             public string Status { get; set; }
